requirements.txt
fastapi
uvicorn
python-multipart
pillow
huggingface_hub
numpy<2
opencv-python-headless
iopath
einops
tqdm
timm
ftfy
regex
pycocotools
matplotlib
scikit-image
pandas
scikit-learn
hydra-core
omegaconf
psutil